@@ -11,6 +11,7 @@ import (
 	"github.com/je4/utils/v2/pkg/zLogger"
 	_ "github.com/lib/pq"
 	"github.com/rs/zerolog"
+	"go.ub.unibas.ch/cloud/certloader/v2/pkg/loader"
 	"io"
 	"log"
 	"os"
@@ -139,12 +140,20 @@ func main() {
 		fair.AddPartition(p)
 	}
 
+	// create TLS Certificate.
+	// the certificate MUST contain <package>.<service> as DNS name
+	serverTLSConfig, serverLoader, err := loader.CreateServerLoader(true, config.TLSConfig, nil, logger)
+	if err != nil {
+		logger.Fatal().Err(err).Msg("cannot create server loader")
+	}
+	defer serverLoader.Close()
+
 	srv, err := service.NewServer(config.ServiceName, config.Addr, config.UserName, config.Password, logger, fair, accessLog, config.JWTKey, config.JWTAlg, config.LinkTokenExp.Duration)
 	if err != nil {
 		logger.Panic().Msgf("cannot initialize server: %v", err)
 	}
 	go func() {
-		if err := srv.ListenAndServe(config.CertPEM, config.KeyPEM); err != nil {
+		if err := srv.ListenAndServe(serverTLSConfig); err != nil {
 			log.Fatalf("server died: %v", err)
 		}
 	}()
